+++ WebUI/languages/es_ES.js
@@ -63,4 +63,4 @@ Language['es_ES'] /* Add/replace the xx_XX here with your language code (like de
 	"": "",
 	"Commands": "",
 	"End Trace": "Finalizar Trazo",
-}
+};